Add missing documentation and improve style for cloud funcitons
