+++ functions/src/commentActions.js
@@ -1,5 +1,6 @@
 const functions = require("firebase-functions");
 const admin = require("firebase-admin");
+const util = require("./util");
 
 function verifyUserMeetsTokenRequirements(userData, postData) {
   // Check that provided user meets required token balance.
@@ -12,18 +13,8 @@ function verifyUserMeetsTokenRequirements(userData, postData) {
   }
 }
 
-async function fetchAndValidateUser(userId) {
-  const userRef = admin.firestore().collection("users").doc(userId);
-  const userDocSnap = await userRef.get();  
-  if (!userDocSnap.exists) {
-    throw new functions.https.HttpsError(
-      "not-found",
-      "Invalid user credentials"
-    );
-  }
-  return {userRef, userDocSnap};
-}
-
+/// Retrieves comment with provided ID from Firestore and returns associated 
+/// document ref and document snapshot.
 async function fetchAndValidateComment(commentId) {
   if (!(typeof commentId === "string" || commentId instanceof String)) {
     // Throwing an HttpsError so that the client gets the error details.
@@ -43,31 +34,12 @@ async function fetchAndValidateComment(commentId) {
   return {commentRef, commentDocSnap};
 }
 
-async function fetchAndValidatePost(postId) {
-  if (!(typeof postId === "string" || postId instanceof String)) {
-    // Throwing an HttpsError so that the client gets the error details.
-    throw new functions.https.HttpsError(
-      "invalid-argument",
-      "Provided Post ID is not a valid string"
-      );
-  }
-  const postRef = admin.firestore().collection("posts").doc(postId);
-  const postDocSnap = await postRef.get();
-  if (!postDocSnap.exists) {
-    throw new functions.https.HttpsError(
-      "not-found",
-      `Provided Post ID: ${postId} does not exist`
-    );
-  }
-  return {postRef, postDocSnap};
-}
-
 /// Add current user to the upvote list of a post.
 exports.upVote = functions.https.onCall(async (data, context) => {
   
   // Get current user
   const uid = context.auth.uid;
-  const {_, userDocSnap} = await fetchAndValidateUser(uid);
+  const {_, userDocSnap} = await util.fetchAndValidateUser(uid);
   
   // Get target comment
   const commentId = data.commentId;
@@ -75,7 +47,7 @@ exports.upVote = functions.https.onCall(async (data, context) => {
 
   // Get target post
   const postId = commentDocSnap.data().postId;
-  const {__, postDocSnap} = await fetchAndValidatePost(postId);
+  const {__, postDocSnap} = await util.fetchAndValidatePost(postId);
 
   // Verify that user has access to the post to which this comment is associated.
   verifyUserMeetsTokenRequirements(userDocSnap.data(), postDocSnap.data());
@@ -103,15 +75,15 @@ exports.downVote = functions.https.onCall(async (data, context) => {
   
   // Get current user
   const uid = context.auth.uid;
-  const {_, userDocSnap} = await fetchAndValidateUser(uid);
+  const {_, userDocSnap} = await util.fetchAndValidateUser(uid);
   
   // Get target comment
   const commentId = data.commentId;
   const {commentRef, commentDocSnap} = await fetchAndValidateComment(commentId);
 
-    // Get target post
-    const postId = commentDocSnap.data().postId;
-    const {__, postDocSnap} = await fetchAndValidatePost(postId);
+  // Get target post
+  const postId = commentDocSnap.data().postId;
+  const {__, postDocSnap} = await util.fetchAndValidatePost(postId);
 
   // Verify that user has access to post.
   verifyUserMeetsTokenRequirements(userDocSnap.data(), postDocSnap.data());
