+++ functions/src/postActions.js
@@ -16,25 +16,6 @@ function verifyUserMeetsTokenRequirements(userData, postData) {
   }
 }
 
-async function fetchAndValidatePost(postId) {
-  if (!(typeof postId === "string" || postId instanceof String)) {
-    // Throwing an HttpsError so that the client gets the error details.
-    throw new functions.https.HttpsError(
-      "invalid-argument",
-      "Provided Post ID is not a valid string"
-    );
-  }
-  const postRef = admin.firestore().collection("postPreviews").doc(postId);
-  const postDocSnap = await postRef.get();
-  if (!postDocSnap.exists) {
-    throw new functions.https.HttpsError(
-      "not-found",
-      `Provided Post ID: ${postId} does not exist`
-    );
-  }
-  return { postRef, postDocSnap };
-}
-
 /// Add current user to the upvote list of a post.
 exports.upVote = functions.https.onCall(async (data, context) => {
   const uid = context.auth.uid;
@@ -44,7 +25,7 @@ exports.upVote = functions.https.onCall(async (data, context) => {
   const { _, userDocSnap } = await util.fetchAndValidateUser(uid);
 
   // Get target post
-  const { postRef, postDocSnap } = await fetchAndValidatePost(postId);
+  const { postRef, postDocSnap } = await util.fetchAndValidatePost(postId);
 
   // Verify that user has access to post.
   verifyUserMeetsTokenRequirements(userDocSnap.data(), postDocSnap.data());
@@ -76,7 +57,7 @@ exports.downVote = functions.https.onCall(async (data, context) => {
   const { _, userDocSnap } = await util.fetchAndValidateUser(uid);
 
   // Get target post
-  const { postRef, postDocSnap } = await fetchAndValidatePost(postId);
+  const { postRef, postDocSnap } = await util.fetchAndValidatePost(postId);
 
   // Verify that user has access to post.
   verifyUserMeetsTokenRequirements(userDocSnap.data(), postDocSnap.data());
